% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_surv.R
\name{gibbs_surv}
\alias{gibbs_surv}
\title{gibbs_surv}
\usage{
gibbs_surv(
  y,
  X,
  nu,
  X_test = NULL,
  nrun = 1000,
  burn = 500,
  thin = 1,
  delta_eta = 0.2,
  delta_omega = 0.2,
  k = NULL,
  verbose = T
)
}
\arguments{
\item{y}{vector of responses}

\item{X}{matrix of predictors}

\item{nu}{vector of censoring indicators}

\item{X_test}{test set matrix (if NULL, this is X)}

\item{nrun}{number of iterations of gibbs sampler}

\item{burn}{number of initial samples to discard}

\item{thin}{thinning amount}

\item{delta_eta}{eta initial acceptance ratio}

\item{delta_omega}{omega initial acceptance ratio}

\item{k}{desired number of factors}

\item{verbose}{description (not sure)}
}
\value{
Matrix (Lambda) of factor loadings
}
\description{
Function to perform factorized regression with the Exponential Survival Model
}
